Do logging direct to journald.
This has the advantage of meaning that the origin process name will be correct
provided it goes through this codepath: fprintf won't work, because if a process
starts children (like e.g. a homescreen does), then their stdout/stderr are
merged into their parent, and journal has no way of differentiating the origin.

We are also able to store information about the context of logging, which might
be useful in post-mortem cases.