MER: Increment loopLevel prior to emitting aboutToBlock()/awake() signals
Without this, any deferred delete actions that occur as a result of
these signals will not be processed until the main event loop exits.

This is triggered by the QtWayland server, which uses aboutToBlock() to
process internal events such as surface destruction.

The patch was not accepted upstream due to insufficient test coverage of the
problem.

Done-with: Aaron Kennedy <aaron.kennedy@jollamobile.com>
Task-number: QTBUG-36434
Change-Id: I3fc143ad442c3a499d09f795355c155ad21b126d