MER: Add libsystrace support.
[MER]: This patch is not upstreamed as libsystrace is not widely in use, the
impact of additional function calls into QtBase was undesired, and the general
approach is not really generic to other types of tracing at this time.

libsystrace is a very small initialized-on-load wrapper for Android systrace
functionality, used so that data is written in order, regardless of how many
different libraries are writing data to it at a given time.

systrace is very useful for tracking system behavior, and for performance
analysis and improvement.

See also:
https://developer.android.com/tools/debugging/systrace.html

Some example uses of QSystraceEvent are also included in this patch. More are
undoubtedly possible. Additional tracing (e.g. counting the number of ongoing
HTTP requests and various other bits and pieces) are also easily accomplished.

Change-Id: I70012f6ddaa084f5b08685ee40795d7fafcd9c96