Remove QT_HAS_CONNECTIONAGENT define.
There is no link-time functionality here, it's purely runtime-detectable (which
we already do anyway), so this define does nothing except to increase
maintenence burden unnecessarily, as the code falls back correctly.

Furthermore, such detection introduces circular dependencies: connectionagent
requires Qt to build, this made Qt require connectionagent to build correctly.

As a bonus, fix compilation of the code inside this define: when linking to
private headers, we must add it to QT, and prefix the header with private/.

Change-Id: Ie7a19baaf1ffbe831066fe2d4208cc3e03905a4d