[qtbase] Fix rectangle shown with hindi text. Fixes JB#32619
After U+0930 U+0947 U+0902 and linefeed, an extra rectangle was rendered
with QML Text element.

Conflicts: tests/auto/gui/text/qtextscriptengine/tst_qtextscriptengine.cpp
-> changes omitted

Original commit 198009db79a85d3cab

Fix several regressions in font selection

In Qt 5.3.0 a change was added which automatically adapts Common
script to surrounding scripts in accordance with the Unicode tr#24.

This broke *a lot* of cases of font selection because the font
selection algorithm is not prepared for handling characters with
adapted scripts. We need to disable this change for now and redo it
later with patches to font selection to avoid the regressions.

[ChangeLog][Text] Fixed several regressions in font selection when
combining different writing systems in the same text.

Task-number: QTBUG-39930
Task-number: QTBUG-39860
Change-Id: Id02b5ae2403c06542ed5d81e7c4deb2e0c7d816e
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>

# Conflicts:
#	tests/auto/gui/text/qtextscriptengine/tst_qtextscriptengine.cpp