Add -no-opengles3 to configure to avoid QT_OPENGL_ES_3 define.
This can be helpful for packaging and when one wants to limit the
features to ES 2.0 to ensure the same code is used on both build
machine and deployment machine. It also prevents inclusion of
GLES3/gl3.h. Windows configure.exe doesn't do auto detection of
GLES3 so it does not get the additional option.

Change-Id: I8750b41635e41bd0ac41d9d59d93570f7d2196e6
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>