[mer][qtnetwork] Create new QNetworkSession from the new online QNetworkConfiguration. Fixes JB#34613
When network configuration configuration changes, use the QNetworkConfiguration
argument within the slot as the QNetworkConfigurationManager::defaultConfiguration()
points to the previous default configuration. Default configration
in turn is wrong e.g. when turning WLAN on so that cellular is
already enabled. This happens because the Qt Connman Bearer plugin finds
the cellular AP from the access point list.

Further, do not close QNetworkSession when changing from WLAN to cellular
as the network session created for WLAN should be kept alive.