[qtbase] Add configuration support for press and hold interval. Contributes to JB#44007
Originally Mer side configuration was implemented with QtQuick2.conf,
but Qt 5.3 added support for platform style on this.

Qt default is 800, but overriding this now with Sailfish value 600
as it's been done on password mask delay already.

When rebasing against new Qt releases, this can be squashed with
"[qtbase] Allow QGenericUnixTheme parameters to be configured."