[qtnetwork] Don't set temp ECDH params for OpenSSL >= 1.1.0. Fixes JB#52403
In OpenSSL versions prior to 1.0.2 default EC temp key params were being
set, but this was also happening for 1.1.0 and beyond, which was
overruling the default auto option. This setting of the parameters has
therefore been removed, as it was preventing some TLS 1.3 handshakes
from completing successfully.

In OpenSSL 1.0.2 support for automatic EC temporary key parameter
selection was introduced, but had to be enabled. In 1.1.0 it became
enabled by default and use of SSL_CTRL_SET_ECDH_AUTO was deprecated.

From the 1.0.2 -> 1.1.0 changelog:

"SSL_{CTX_}set_ecdh_auto() has been removed and ECDH is support is always
enabled now. If you want to disable the support you should exclude it
using the list of supported ciphers. This also means that the
"-no_ecdhe" option has been removed from s_server."