[qtbase] Fix QDateTime::toString() timezone info. Contributes to JB#53010
Prior to commit 7c33c644 upstream, the time zone offset is not appended
to string generated by toString with the Qt::ISODate format. The present
commit is a rewrite of the upstream commit targetting 5.6. The present
commit can be safely ignored when moving to 5.9 or more recent versions
of Qt.

It is somehow related to QT-26161.