[qtnetwork] Remove SSL verification error locks
This commit should solve QTBUG-76157 by removing the mutex-based error
list and instead attaching the error list to either the X509_STORE or
SSL struct as ex_data.

This bug has been fixed in upstream Qt since 5.12, so this commit may
be ignored when upgrading to that Qt version or higher.